Initial commit
Basic working implementation of a solver for the incompressible
Navier-Stokes equations on a MAC grid. See the folder ns/ for relevant
code. The solver uses a projection method (PmIII, see Brown, Cortez,
and Minion, 2001). This requires a linear solve of a Poisson equation,
which is done by multigrid. See the folder mg/ for the relevant code.
Time stepping uses a 2-stage RK scheme.

Additionally, The ib/ folder contains classes useful for interacting
with Lagrangian grids. The transfer operator exists as
`ib::grid::interpolate`. These grids are also vectors capable of being
used as velocity components within the Navier-Stokes solver.