Remove templates to allow for easier optimizations
The old code used a lot of templates everywhere. That's pretty annoying
to deal with when we want to apply optimizations to a particular set of
operations.

These changes remove those templates and allow us to use BLAS routines
in place of some optimizations (only some MV multiplies are currently
optimized). This should not affect the speed of the code, for the most
part (kdtrees and multigrid solvers now use one more level of pointer
redirection).

The code now comes with a proper build system, just type `make'. To run
the tests, type `make test'. To clean up, `make clean'. And to *really*
clean up, `make distclean'.